# backend/docs/swagger.yaml
definitions:
  diploma_modules_auth_handler_model.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  diploma_modules_cart_handler_model.CheckoutRequest:
    properties:
      address_id:
        type: integer
    required:
    - address_id
    type: object
  diploma_modules_cart_handler_model.CheckoutResponse:
    properties:
      checkout_url:
        type: string
      order_id:
        type: string
      payment_id:
        type: string
    type: object
  diploma_modules_cart_handler_model.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  diploma_modules_cart_handler_model.MessageResponse:
    properties:
      message:
        type: string
    type: object
...
  diploma_modules_product_handler_model.DetailedProduct:
    properties:
      product:
        $ref: '#/definitions/diploma_modules_product_handler_model.Product'
      reviews:
        items:
          $ref: '#/definitions/diploma_modules_product_handler_model.Review'
        type: array
      suppliers:
        items:
          $ref: '#/definitions/diploma_modules_product_handler_model.ProductSupplier'
        type: array
    type: object
  diploma_modules_product_handler_model.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  diploma_modules_product_handler_model.Product:
    properties:
      id:
        type: integer
      image:
        type: string
      lowest_product_supplier:
        $ref: '#/definitions/diploma_modules_product_handler_model.ProductSupplier'
      name:
        type: string
      rating_average:
        type: number
      rating_count:
        type: integer
    type: object
  diploma_modules_product_handler_model.ProductSupplier:
    properties:
      price:
        type: integer
      sell_amount:
        type: integer
      supplier:
        $ref: '#/definitions/diploma_modules_product_handler_model.Supplier'
    type: object
  diploma_modules_product_handler_model.Review:
    properties:
      comment:
        type: string
      created_at:
        type: string
      id:
        type: integer
      rating:
        type: integer
      user_id:
        type: integer
      user_name:
        type: string
    type: object
...
  diploma_modules_product_handler_model.Supplier:
    properties:
      delivery_fee:
        type: integer
      free_delivery_amount:
        type: integer
      id:
        type: integer
      name:
        type: string
      order_amount:
        type: integer
    type: object
  diploma_modules_user_handler_model.Address:
    properties:
      description:
        type: string
      id:
        type: integer
      street:
        type: string
    type: object
  diploma_modules_user_handler_model.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  model.ContractResponse:
    properties:
      content:
        type: string
      customer_signature:
        type: string
      id:
        type: integer
      status:
        type: integer
      supplier_signature:
        type: string
    type: object
...
  model.GetAddressResponse:
    properties:
      address_list:
        items:
          $ref: '#/definitions/diploma_modules_user_handler_model.Address'
        type: array
    type: object
...
  model.LoginInput:
    properties:
      password:
        example: secure123
        type: string
      phone_number:
        example: "+123456789"
        type: string
    type: object
  model.LoginResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
...
  model.ProductResponse:
    properties:
      product:
        $ref: '#/definitions/diploma_modules_product_handler_model.DetailedProduct'
    type: object
...
  model.RegisterInput:
    properties:
      confirm_password:
        example: secure123
        type: string
      name:
        example: John Doe
        type: string
      password:
        example: secure123
        type: string
      phone_number:
        example: "+123456789"
        type: string
    type: object
  model.RegisterResponse:
    properties:
      id:
        example: 1
        type: integer
    type: object
  model.SetAddressInput:
    properties:
      address:
        $ref: '#/definitions/diploma_modules_user_handler_model.Address'
    type: object
  model.SignRequest:
    properties:
      contract_id:
        type: integer
      signature:
        type: string
    required:
    - contract_id
    - signature
    type: object
  model.StatusResponse:
    properties:
      status:
        type: string
    type: object
...
info:
  contact: {}
  description: This is a sample server with JWT authorization.
  title: Go JWT Swagger Example API
  version: "1.0"
paths:
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Login user and return tokens
      parameters:
      - description: Login input
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.LoginResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/diploma_modules_auth_handler_model.ErrorResponse'
      summary: User login
      tags:
      - auth
  /api/auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user
      parameters:
      - description: Register input
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.RegisterInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.RegisterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/diploma_modules_auth_handler_model.ErrorResponse'
      summary: User registration
      tags:
      - auth
  /api/callback/checkout:
    post:
      consumes:
      - application/json
      description: Receive payment provider callback and finalize checkout
      parameters:
      - description: Payment callback payload
        in: body
        name: payload
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/diploma_modules_cart_handler_model.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/diploma_modules_cart_handler_model.ErrorResponse'
      summary: Process cart payment callback
      tags:
      - cart
...
  /api/cart/checkout:
    post:
      consumes:
      - application/json
      description: Processes the checkout of the authenticated user's cart.
      parameters:
      - description: Checkout payload
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/diploma_modules_cart_handler_model.CheckoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Checkout status
          schema:
            $ref: '#/definitions/diploma_modules_cart_handler_model.CheckoutResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/diploma_modules_cart_handler_model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/diploma_modules_cart_handler_model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/diploma_modules_cart_handler_model.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Process checkout operation
      tags:
      - cart
...
  /api/cart/mock-success:
    get:
      description: Debug endpoint that simulates successful payment for an order
      parameters:
      - description: Order ID
        in: query
        name: order_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "302":
          description: Found
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/diploma_modules_cart_handler_model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/diploma_modules_cart_handler_model.ErrorResponse'
      summary: Mock successful checkout
      tags:
      - cart
  /api/contract:
    get:
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.ContractResponse'
            type: array
      security:
      - ApiKeyAuth: []
      summary: List of contracts for the current user
      tags:
      - contracts
  /api/contract/{id}:
    get:
      consumes:
      - application/json
      description: Returns the contract by ID
      parameters:
      - description: Contract ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Contract
          schema:
            $ref: '#/definitions/model.ContractResponse'
        "400":
          description: Invalid ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get contract
      tags:
      - contracts
  /api/contract/sign:
    post:
      consumes:
      - application/json
      description: Signing the contract (by client or supplier)
      parameters:
      - description: Contract ID and Signature
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.SignRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Signature saved
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Validation error
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Sign the contract
      tags:
      - contracts
...
  /api/product/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve product information by its ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ProductResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/diploma_modules_product_handler_model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/diploma_modules_product_handler_model.ErrorResponse'
      summary: Get product by ID
      tags:
      - product
...
  /api/user/address:
    get:
      description: Retrieve address list for a user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.GetAddressResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/diploma_modules_user_handler_model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/diploma_modules_user_handler_model.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get address
      tags:
      - user
    post:
      consumes:
      - application/json
      description: --
      parameters:
      - description: something
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.SetAddressInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.StatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/diploma_modules_user_handler_model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/diploma_modules_user_handler_model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/diploma_modules_user_handler_model.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Set address
      tags:
      - user
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"